Moving last things from /tests, removing dead code
