+++ ecs/e2e/ecs-local/context_test.go
@@ -23,7 +23,7 @@ import (
 
 	"gotest.tools/v3/icmd"
 
-	. "github.com/docker/compose-cli/tests/framework"
+	. "github.com/docker/compose-cli/utils/e2e"
 )
 
 const (
