+++ local/e2e/cli-only/mockmetrics.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package framework
+package main
 
 import (
 	"io/ioutil"
