+++ local/e2e/compose/logs_test.go
@@ -24,7 +24,7 @@ import (
 
 	"gotest.tools/v3/icmd"
 
-	. "github.com/docker/compose-cli/tests/framework"
+	. "github.com/docker/compose-cli/utils/e2e"
 )
 
 func TestLocalComposeLogs(t *testing.T) {
