--- tests/composefiles/nginx.yaml
@@ -1,5 +1,0 @@
-services:
-  nginx:
-    image: nginx
-    ports:
-      - "80:80"
