--- tests/framework/unit.go
@@ -1,42 +1,0 @@
-/*
-   Copyright 2020 Docker Compose CLI authors
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package framework
-
-import (
-	"context"
-	"io/ioutil"
-	"os"
-)
-
-// TestCLI is a helper struct for CLI tests.
-type TestCLI struct {
-	ctx    context.Context
-	writer *os.File
-	reader *os.File
-}
-
-// Context returns a configured context
-func (c *TestCLI) Context() context.Context {
-	return c.ctx
-}
-
-// GetStdOut returns the output of the command
-func (c *TestCLI) GetStdOut() string {
-	_ = c.writer.Close()
-	out, _ := ioutil.ReadAll(c.reader)
-	return string(out)
-}
